// tutor/src/payload-types.ts
export interface Course {
  id: string;
  courseName: string;
  tuition?: number | null;
  total_hours?: number | null;
  skills?: ('reading' | 'writing' | 'speaking' | 'listening' | 'grammar' | 'vocabulary')[] | null;
  type?: ('online' | 'offline') | null;
  thumbnail?: (string | null) | Media;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt?: string | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Admin {
  id: string;
  department: 'ceo' | 'hoc_vu' | 'hoc_thuat' | 'sale' | 'marketing' | 'hr' | 'ke_toan' | '25';
  role?:
    | (
        | 'admin'
        | 'hoc_vu_manager'
        | 'hoc_vu_executive'
        | 'academic_manager'
        | 'academic_executive'
        | 'teacher_part_time'
        | 'teacher_full_time'
        | 'wow_manager'
        | 'wow_executive'
        | 'sale_manager'
        | 'sale_executive'
        | 'marketing_manager'
        | 'marketing_executive'
        | 'designer'
        | 'seo'
        | 'digital_marketing'
        | 'hr_manager'
        | 'hr_generalist'
        | 'ke_toan'
      )
    | null;
  fullName: string;
  phone: string;
  birthday?: string | null;
  working_date?: string | null;
  address?: string | null;
  avatar?: (string | null) | Media;
  /**
   * Dùng để gửi cho học viên khi học WOW trực tuyến
   */
  meetingLink?: string | null;
  availability?:
    | {
        /**
         * Chọn ngày rảnh của WOW để học sinh có thể book. Ví dụ bắt đầu vào ngày 01/01/2025 9:00 và kết thúc vào ngày 01/01/2025 17:00. Có thể chọn nhiều khung giờ khác nhau cùng ngày nếu như ngày đó WOW rảnh vào nhiều khung giờ
         */
        start_date: string;
        end_date: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  email?: string | null;
  username: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'admins';
}
// ...
export interface Lead {
  id: string;
  displayTitle?: string | null;
  title?: string | null;
  fullName?: string | null;
  phone?: string | null;
  email?: string | null;
  gender?: ('male' | 'female') | null;
  'parent name'?: string | null;
  'parent phone'?: string | null;
  'parent email'?: string | null;
  date_of_birth?: string | null;
  link_facebook?: string | null;
  address?: string | null;
  type?: ('student' | 'college_student' | 'office_worker' | 'parent' | 'other') | null;
  target?: number | null;
  wow_quota?: number | null;
  purpose?: string | null;
  branch?: (string | null) | Branch;
  source:
    | 'facebook'
    | 'google'
    | 'hotline'
    | 'walking'
    | 'referral'
    | 'seeding'
    | 'school_tour'
    | 'sponsorship'
    | 're_enroll'
    | 'from_website'
    | 'other';
  saleInCharge?: (string | null) | Admin;
  status?:
    | (
        | 'waiting_for_call'
        | 'test_booked'
        | 'test_not_booked'
        | 'potential'
        | 'not_potential'
        | 'paid_case'
        | 'failed_case'
        | 'consider'
      )
    | null;
  courses?: string | null;
  coupon?: (string | Coupon)[] | null;
  notes?:
    | {
        channel_interaction?: ('zalo' | 'phone' | 'fanpage' | 'hotline' | 'direct_message' | 'online') | null;
        appointment_date?: string | null;
        status?:
          | (
              | 'waiting_for_call'
              | 'test_booked'
              | 'test_not_booked'
              | 'potential'
              | 'not_potential'
              | 'paid_case'
              | 'failed_case'
              | 'consider'
            )
          | null;
        message?: string | null;
        createdAt?: string | null;
        createdBy?: (string | null) | Admin;
        id?: string | null;
      }[]
    | null;
  chats?:
    | {
        createdBy?: (string | null) | Admin;
        message?: string | null;
        createdAt?: string | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "branches".
 */
export interface Branch {
  id: string;
  name: string;
  address: string;
  phone: string;
  rooms: (string | Room)[];
  status?: ('active' | 'unactive') | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "rooms".
 */
export interface Room {
  id: string;
  name: string;
  capacity?: number | null;
  status?: ('active' | 'unactive') | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "coupons".
 */
export interface Coupon {
  id: string;
  name: string;
  discount: number;
  type: 'money' | 'percent';
  startDate: string;
  endDate: string;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface User {
  id: string;
  avatar?: (string | null) | Media;
  role: 'user';
  lead?: (string | null) | Lead;
  violationStatus?: ('none' | 'warning_1' | 'warning_2' | 'terminated') | null;
  displayTitle?: string | null;
  updatedAt: string;
  createdAt: string;
  email?: string | null;
  username: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "orders".
 */
export interface Order {
  id: string;
  customer?: (string | null) | User;
  courses?: (string | Course)[] | null;
  coupon?: (string | Coupon)[] | null;
  status?: ('full' | 'installment') | null;
  note?: string | null;
  saleInCharge?: string | null;
  updatedAt: string;
  createdAt: string;
}